Refactor model classes: update Client, Order, OrderItem, and Product to use Double for monetary values, add OrderStatus enum, and enhance relationships with proper annotations.

+++ src/main/java/com/ismail/smartShop/model/Product.java
@@ -27,7 +27,7 @@ public class Product {
 
     private String nom;
 
-    private  BigDecimal  prixUnit;
+    private  Double  prixUnit;
 
     private Integer stockQuantitie;
